fix: update market_cap schema to float to support large numbers
Results Summary:
Successfully tested fetching stock quotes for TCS and INFY with the updated schema. Market capitalization values are now correctly parsed and returned as floats, preventing validation errors that occur when using integers.

Results:
- TCS Market Cap: 8473560967156.0
- TCS Price: 2342.0
- INFY Market Cap: 4633341373232.0
- INFY Price: 1144.0

+++ app/schemas.py
@@ -387,7 +387,7 @@ class StockQuote(BaseModel):
     change: float = Field(default=0.0, description="Absolute price change.")
     change_percent: float = Field(default=0.0, description="Percentage price change.")
     volume: int = Field(default=0, description="Trading volume.")
-    market_cap: Optional[int] = Field(default=None, description="Market capitalization.")
+    market_cap: Optional[float] = Field(default=None, description="Market capitalization.")
     day_open: Optional[float] = Field(default=None, description="Day open price.")
     day_high: Optional[float] = Field(default=None, description="Day high price.")
     day_low: Optional[float] = Field(default=None, description="Day low price.")
